moved validator to resource
